added joink, work on checking for higher rank set
Some renaming of functions for consistency and better mnemonics
Added a joink verb that joins dictionaries, either on the fly, or not,
along the keys
Started work on joinv, that will join the values, keeping keys the same.
Not complete yet.
Started work on input rank validation for dyadic set, where previously
all sorts of errors and problems could come up because of length
mismatch, resulting in an inconsistent dictionary state.